+++ api/src/main/java/uk/gov/hmcts/payment/api/controllers/CardPaymentController.java
@@ -31,9 +31,9 @@ import uk.gov.hmcts.payment.api.service.FeePayApportionService;
 import uk.gov.hmcts.payment.api.service.PciPalPaymentService;
 import uk.gov.hmcts.payment.api.v1.model.exceptions.PaymentException;
 import uk.gov.hmcts.payment.api.v1.model.exceptions.PaymentNotFoundException;
+import uk.gov.hmcts.payment.api.validators.PaymentValidator;
 
 import javax.validation.Valid;
-import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -59,6 +59,7 @@ public class CardPaymentController {
     private final FF4j ff4j;
     private final FeePayApportionService feePayApportionService;
     private final LaunchDarklyFeatureToggler featureToggler;
+    private final PaymentValidator paymentValidator;
 
     @Autowired
     public CardPaymentController(DelegatingPaymentService<PaymentFeeLink, String> cardDelegatingPaymentService,
@@ -66,7 +67,8 @@ public class CardPaymentController {
                                  CardDetailsService<CardDetails, String> cardDetailsService,
                                  PciPalPaymentService pciPalPaymentService,
                                  FF4j ff4j,
-                                 FeePayApportionService feePayApportionService,LaunchDarklyFeatureToggler featureToggler) {
+                                 FeePayApportionService feePayApportionService,LaunchDarklyFeatureToggler featureToggler,
+                                 PaymentValidator paymentValidator) {
         this.delegatingPaymentService = cardDelegatingPaymentService;
         this.paymentDtoMapper = paymentDtoMapper;
         this.cardDetailsService = cardDetailsService;
@@ -74,6 +76,7 @@ public class CardPaymentController {
         this.ff4j = ff4j;
         this.feePayApportionService = feePayApportionService;
         this.featureToggler = featureToggler;
+        this.paymentValidator = paymentValidator;
     }
 
     @ApiOperation(value = "Create card payment", notes = "Create card payment")
@@ -90,7 +93,7 @@ public class CardPaymentController {
         @RequestHeader(value = "service-callback-url", required = false) String serviceCallbackUrl,
         @Valid @RequestBody CardPaymentRequest request) throws CheckDigitException, URISyntaxException {
 
-        if(!validateReturnUrl(returnURL)) {
+        if(!paymentValidator.validateReturnUrl(returnURL)) {
             return new ResponseEntity("returnUrl: Must be an internal domain of hmcts.net or gov.uk", BAD_REQUEST);
         }
 
@@ -236,25 +239,4 @@ public class CardPaymentController {
     public String return403(AccessDeniedException ex) {
         return ex.getMessage();
     }
-
-    private boolean validateReturnUrl(String returnUrl) throws URISyntaxException {
-        if(returnUrl != null) {
-            String hostName = getHostName(returnUrl);
-            if(StringUtils.isNotEmpty(hostName) && (hostName.endsWith("hmcts.net") || hostName.endsWith("gov.uk"))){
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private String getHostName(String url) throws URISyntaxException {
-        URI uri = new URI(url);
-        String hostname = uri.getHost();
-        // to provide faultproof result, check if not null then return only hostname, without www.
-        if (hostname != null) {
-            return hostname.startsWith("www.") ? hostname.substring(4) : hostname;
-        }
-        return hostname;
-    }
-
 }
